refactor: Splitting MyApp into separate files
